Code cleanup / new files PART3
