+++ HowManyCities/Models/MapGuessModel.swift
@@ -1,0 +1,41 @@
+//
+//  MapGuessModel.swift
+//  HowManyCities
+//
+//  Created by Geoffrey Liu on 4/23/22.
+//
+
+import Foundation
+import MapKit
+
+struct MapGuessModel: Codable {
+  var guessedCities: Set<City> = .init()
+  var gameConfiguration: GameConfiguration?
+  var startTime: Date = .now
+  var usedMultiCityInput: Bool = false
+  var lastRegion: MKCoordinateRegion = .init(center: .zero, span: .full)
+  
+  mutating func resetState() {
+    guessedCities = .init()
+  }
+}
+
+enum CityGuessError: Error {
+  case noneFound(_ cityName: String)
+  case alreadyGuessed
+  case emptyGuess
+  case serverError
+  
+  var message: String {
+    switch self {
+      case .noneFound(let name):
+        return "No city named \(name.capitalized) found!"
+      case .alreadyGuessed:
+        return "Already guessed!"
+      case .emptyGuess:
+        return "Please enter a city name"
+      case .serverError:
+        return "We're having technical issues, please try again"
+    }
+  }
+}
